Run against the BeiDou server: v95 packet layouts and unsupported opcodes
* the handler table reaches opcode 0x1000 (the server's HP/MP alert packet), and
  opcodes the client does not act on get layout parsers (Net/UnsupportedPackets,
  Net/Handlers/Unsupported) so a wire change shows up as a parse error instead of
  silence
* char list, inventory, quest, skill, item, cash shop, login and movement parsers
  follow the server's layouts (fourth job master levels, quest info entries, Evan SP
  per job, channel load scaled to 800, ...)
* single-package .nx sets are accepted as well (USE_NX_V83), so the data of a
  pre-split client runs as it is
* buffs and the tooltips take the fields those packets carry

+++ Net/Handlers/InventoryHandlers.cpp
@@ -62,9 +62,12 @@ namespace ms
 
 		std::vector<Mod> mods;
 
-		int8_t size = recv.read_byte();
+		// The server writes the mod count as one truncated byte (PacketCreator.java:2448,
+		// ByteBufOutPacket.java:48-49), so it must be read unsigned: a signed read turns
+		// counts of 128-255 into negative values and drops every mod of the packet.
+		uint8_t size = static_cast<uint8_t>(recv.read_byte());
 
-		for (int8_t i = 0; i < size; i++)
+		for (uint8_t i = 0; i < size; i++)
 		{
 			Mod mod;
 			mod.mode = recv.read_byte();
